[Feature #161209936] move config file
