+++ app/__init__.py
@@ -1,7 +1,7 @@
 import os
 from flask import Flask
 from flask_restful import Api
-from instance.config import app_config
+from config import app_config
 
 from .jwt_instance import jwt
 from .api.v1 import apiv1
@@ -14,7 +14,6 @@ def create_app(config_name):
     app = Flask(__name__, instance_relative_config=True)
 
     app.config.from_object(app_config[config_name])
-    app.config.from_pyfile('config.py')
 
     jwt.init_app(app)
