+++ app/api/v1/models/category.py
@@ -1,7 +1,7 @@
 from .base import BaseModel
 
 """
-Category Model
+Category Model has saved data
 
 """
