+++ app/jwt_instance.py
@@ -1,8 +1,16 @@
 from flask_jwt_extended import JWTManager
 
 from .api.v1.models.auth import UserModel
-
+"""
+     Holds the the JWT instance to be used in the app
+"""
 jwt = JWTManager()
+
+"""
+    Function handles revoked tokens once a user is logged out 
+    It checks if token is blacklisted.
+
+"""
 
 
 @jwt.token_in_blacklist_loader
