+++ run.py
@@ -16,4 +16,3 @@ def home():
 # run the app
 if __name__ == '__main__':
     app.run()
-
